refactor: LLM 서비스 로직 리펙토링
